Add tabulate package to represent cubes as strings.
Require the tabulate package. Reformat cube __str__ related code to
make use of it.

One thing which was done in this commit that may seem strange is the
assertions in the __str__ method of base_cube. These were added to
make headers more explicit for future programmers but they also serve
as cheap unit tests. They aren't necessary and could safely be
removed. The headers could be squashed into the table also, and with
headers="firstrow". That way one would not need to return multiple
values.

Resolves #1