Fix descriptor
4d8cd8ed9752b3f98bd1147eb97fa267ceb5225f attempted to introduce
a per object instance descriptor in an attempt to avoid potential
multi-threading issues with WeakRefDictionary. Instead, it
created a single descriptor per type, which ended up being shared
by instances of that type.

This change reintroduces the WeakRefDictionary implementation.
Due to the GIL python data structures are inherently thread-safe,
for the moment.