Add array_extents and 'single' boolean arg
Add a 'single' boolean keyword argument to _dim_attribute.
If True, single element lists will be converted to a single object.
If False, the list will be returned.

Also added an array_extents method that, given an array name,
returns a list of (lower, upper) extent tuples for the array schema.

Also added an array_slice_index method that, given an array name,
returns a tuple of slice(lower, upper, 1) objects for the array schema.